backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.34
psycopg2-binary==2.9.9
alembic==1.13.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.8.2
pydantic-settings==2.4.0
python-multipart==0.0.9
email-validator==2.2.0
boto3==1.34.148
celery==5.4.0
redis==5.0.8
httpx==0.27.2
reportlab==4.2.2
pdfminer.six==20240706
pillow==10.4.0
pytesseract==0.3.13
tenacity==9.0.0
